requirements.txt
requests==2.32.0
pandas==1.4.2
python-decouple==3.6
